+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 * Added Language System
 * Added Serialization runtime and Coroutine Dependencies
 * Cleaned up dependencies (by excluding duplicated depends)
+### 1.2
+* Added Arrow Core
+* Added Command Manager [WIP]
+* Redesigned Language System
+* Added Server Data Manager